Only call line_item.reload if the item is persisted when finalizing
Reloading line_items on a persisted order is the right thing to do before
calculating taxes on a persisted order. If the order is not persisted, this will
clear out all of the line items on the order. In these cases, we should not
attempt to reload the line items.