Update lines to fallback to object_id on non-persisted items
When creating the request to send to avatax for tax calculation, the ids of
shipments and line items are used as a unique identifier. With a non-persisted
line_item or shipment, the id will be nil, leading to an avatax error about
duplicate line numbers if there are multiple shipments or line items.

Additionally, the response from avatax is cached and reading from the response
for the items under consideration is necessary.
Therefore, we need to use a unique identifier for each line_item and shipment
when they are not persisted and have no set id.

Since the work to calculate these taxes is done in a single request, utilizing
the object_id of the line_items and shipments allows us to produce a unique line
number for each item and shipment that can be retrieved with the same id from
the cache later in the request.

If the order is persisted, the cache key used to save the avatax response is
changed, and the regular process of relying on the item.id will work as
expected.