Use `Variant#sku` in line item cache key for non-persisted records
This change improves the `avatax_cache_key` for line items which are not
persisted by using the variant SKU instead of the `id` which would be
nil. This will ensure that two people with the same address but
different items in their cart (with the same quantity and price)
do not get the same tax rate incorrectly. This is an edge case, however
it may happen on stores where items with the same price can be in
different tax categories.