Use `Address#avatax_cache_key` instead of default `cache_key`
This change uses the newly added `avatax_cache_key` on addresses instead
of the stock Rails one. This is necessary so we can correctly take into
account user's addresses when caching the response for in-memory orders.
This change fixes a bug which would cause users to see a wrong tax
amount due to a bad cache hit from other users who have the same number
of items and the same priced items, but for a different address.